change from int to float

+++ src/dataloaders/yamada.py
@@ -33,12 +33,12 @@ class YamadaPershina(object):
         result = []
 
         # Each abstract is of shape num_ents * NUMBER_CANDIDATES
-        all_candidates = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.int64)
-        labels = np.zeros(self.args.max_ent_size).astype(np.int64)
+        all_candidates = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.int32)
+        labels = np.zeros(self.args.max_ent_size).astype(np.int32)
 
         if self.args.include_string:
-            exact_match = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.int64)
-            contains = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.int64)
+            exact_match = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.float32)
+            contains = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.float32)
         if self.args.include_stats:
             candidate_priors = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.float32)
             candidate_conditionals = np.zeros((self.args.max_ent_size, self.number_candidates)).astype(np.float32)
